Changing troop id from byte to ushort. Making it so you need at least 40 upkeep to join an assignment.

+++ Game/Data/Troop/TroopTemplate.cs
@@ -71,7 +71,7 @@ namespace Game.Data.Troop
                 return new[]
                 {
                         new DbColumn("city_id", stub.City.Id, DbType.UInt32),
-                        new DbColumn("troop_stub_id", stub.TroopId, DbType.UInt32),
+                        new DbColumn("troop_stub_id", stub.TroopId, DbType.UInt16),
                 };
             }
         }
